+++ iOS/3.4/pubnub/libs/PubNub/Data/PNConfiguration.m
@@ -213,7 +213,7 @@
                                                                                      kPNServiceMainDomain]
                                                                          withString:@""];
 
-        self.origin = [NSString stringWithFormat:@"%@-%d.%@", subDomain, PNRandomInteger(), kPNServiceMainDomain];
+        self.origin = [NSString stringWithFormat:@"%@-%ld.%@", subDomain, (long)PNRandomInteger(), kPNServiceMainDomain];
     }
     else {
